+++ et-driver/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+- [SW-16696] Consistently use devnum to represent device number and provide devnum through ioctl
 ### Deprecated
 ### Removed
 ### Fixed